ajout affichage et suppression commentaires

+++ app/models/commentary_model.py
@@ -1,11 +1,12 @@
 from ..database import get_db
+from datetime import datetime
 
 def create_commentary(message, post_id, user_id):
     conn = get_db()
     cur = conn.cursor()
 
     try:
-        cur.execute("INSERT INTO Post (messsage, post_id, user_id) VALUES (?,?,?)", (message, post_id, user_id))
+        cur.execute("INSERT INTO Commentary (commentary, post_id, user_id, creation_date) VALUES (?,?,?,?)", (message, post_id, user_id, datetime.now()))
         conn.commit()
     except Exception: 
         raise
@@ -13,12 +14,12 @@ def create_commentary(message, post_id, user_id):
         cur.close()
         conn.close()
 
-def delete_commentary(commentary_id):
+def delete_commentary(commentary_id, user_id):
     conn = get_db()
     cur = conn.cursor()
 
     try:
-        cur.execute("DELETE FROM commentary WHERE commentary_id = ?", (commentary_id,))
+        cur.execute("DELETE FROM commentary WHERE commentary_id=? AND user_id=?", (commentary_id, user_id))
         conn.commit()
     except Exception:
         raise
@@ -33,6 +34,20 @@ def update_commentary(commentary_id, message):
     try:
         cur.execute("UPDATE commentary SET message = ? WHERE commentary_id = ?", (message, commentary_id))
         conn.commit()
+    except Exception:
+        raise
+    finally:
+        cur.close()
+        conn.close()
+
+def get_commentaries(post_id):
+    conn = get_db()
+    cur = conn.cursor()
+
+    try:
+        cur.execute("SELECT c.commentary_id, c.commentary, u.first_name, u.last_name, c.user_id FROM Commentary c JOIN User u ON u.user_id=c.user_id WHERE c.post_id=?", (post_id,))
+        commentaries = cur.fetchall()
+        return commentaries
     except Exception:
         raise
     finally:
